messageBoard: rename overviewView -> messageBoardView
