+++ webapp/app/landing/views/messageBoard/index.js
@@ -1,1 +1,1 @@
-export { default } from './overviewView'
+export { default } from './messageBoardView'
